ev - added controllers for home and eartquakes and fixed imports and tests
