+++ src/test/java/hello/HomePageTest.java
@@ -13,6 +13,9 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+
+import hello.controllers.HomeController;
+
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.xpath;
 
 import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
@@ -24,7 +27,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 
 @RunWith(SpringRunner.class)
 
-@WebMvcTest(WebController.class)
+@WebMvcTest(HomeController.class)
 public class HomePageTest {
 
     @Autowired
